Moved the state management logic for menu and expanded groups to custom hooks
